+++ plugins/healthcheck/plugin.ts
@@ -1,11 +1,13 @@
 import { Api, EMethod, Route } from "../../mod.ts";
 
+interface IConfig {
+  uri?: string
+}
+
 /**
- * @example plugin(api, { log: console.log })
+ * get status 200 for this healthcheck endpoint
+ * @example plugin(api)
  **/
-export default function plugin(api: Api, config: IConfig) {
-  /**
-   * preset healthcheck endpoint
-   */
-  api.addRoute(new Route([EMethod.HEAD, EMethod.GET], "/healthz"));
+export default function plugin(api: Api, config: IConfig = {}) {
+  api.addRoute(new Route([EMethod.HEAD, EMethod.GET], config?.uri ?? "/healthz"));
 }