+++ plugins/access-log/plugin.test.ts
@@ -1,4 +1,4 @@
-import { assertEquals } from "../../src/dev_deps.ts";
+import {assertCalledCount} from "../../src/dev_deps.ts";
 import { mockApi, mockFn } from "../../dev_mod.ts";
 import { Route } from "../../mod.ts";
 import plugin from "./plugin.ts";
@@ -14,8 +14,9 @@ Deno.test("Access log plugin", async () => {
   // install plugin
   plugin(api, { log });
 
-  
+  assertCalledCount(log, 0);
 
   await api.sendByArguments("GET", "/hello");
 
+  assertCalledCount(log, 1);
 });